+++ Sources/Sentry/SentrySDK.m
@@ -184,14 +184,37 @@ static NSUInteger startInvocations;
 
 + (id<SentrySpan>)startTransactionWithName:(NSString *)name operation:(NSString *)operation
 {
-    return [SentrySDK.currentHub startTransactionWithName:name operation:operation];
+    return [self startTransactionWithName:name
+                               nameSource:kSentryTransactionNameSourceCustom
+                                operation:operation];
+}
+
++ (id<SentrySpan>)startTransactionWithName:(NSString *)name
+                                nameSource:(SentryTransactionNameSource)source
+                                 operation:(NSString *)operation
+{
+    return [SentrySDK.currentHub startTransactionWithName:name
+                                               nameSource:source
+                                                operation:operation];
 }
 
 + (id<SentrySpan>)startTransactionWithName:(NSString *)name
                                  operation:(NSString *)operation
                                bindToScope:(BOOL)bindToScope
 {
+    return [self startTransactionWithName:name
+                               nameSource:kSentryTransactionNameSourceCustom
+                                operation:operation
+                              bindToScope:bindToScope];
+}
+
++ (id<SentrySpan>)startTransactionWithName:(NSString *)name
+                                nameSource:(SentryTransactionNameSource)source
+                                 operation:(NSString *)operation
+                               bindToScope:(BOOL)bindToScope
+{
     return [SentrySDK.currentHub startTransactionWithName:name
+                                               nameSource:source
                                                 operation:operation
                                               bindToScope:bindToScope];
 }