+++ Sources/Sentry/SentryTracer.m
@@ -145,7 +145,7 @@ static NSLock *profilerLock;
 {
     if (self = [super init]) {
         self.rootSpan = [[SentrySpan alloc] initWithTransaction:self context:transactionContext];
-        self.name = transactionContext.name;
+        self.transactionContext = transactionContext;
         _children = [[NSMutableArray alloc] init];
         self.hub = hub;
         self.isWaitingForChildren = NO;
@@ -565,7 +565,7 @@ static NSLock *profilerLock;
     }
 
     SentryTransaction *transaction = [[SentryTransaction alloc] initWithTrace:self children:spans];
-    transaction.transaction = self.name;
+    transaction.transaction = self.transactionContext.name;
     [self addMeasurements:transaction appStartMeasurement:appStartMeasurement];
     return transaction;
 }